+++ pkg/graph/store/memory/capabilities.go
@@ -1,6 +1,7 @@
 package memory
 
 import (
+	"gitlab2024.bds421-cloud.com/bds421/rho/tkg/v3/pkg/graph/internal/generatedcreate"
 	storecontract "gitlab2024.bds421-cloud.com/bds421/rho/tkg/v3/pkg/graph/store"
 )
 
@@ -9,21 +10,22 @@ import (
 // or split, the affected line fails to compile here, surfacing the
 // adaptation cost at the implementation boundary instead of at call sites.
 var (
-	_ storecontract.Store                           = (*Store)(nil)
-	_ storecontract.Lifecycle                       = (*Store)(nil)
-	_ storecontract.NodeCRUDCapability              = (*Store)(nil)
-	_ storecontract.RelationshipCRUDCapability      = (*Store)(nil)
-	_ storecontract.AdjacencyCapability             = (*Store)(nil)
-	_ storecontract.BulkReadCapability              = (*Store)(nil)
-	_ storecontract.BatchCapability                 = (*Store)(nil)
-	_ storecontract.HistoryCapability               = (*Store)(nil)
-	_ storecontract.StatsCapability                 = (*Store)(nil)
-	_ storecontract.IterationCapability             = (*Store)(nil)
-	_ storecontract.NodeIntegrityHashCapability     = (*Store)(nil)
-	_ storecontract.EndpointIntegrityHashCapability = (*Store)(nil)
-	_ storecontract.PropertyIndexCapability         = (*Store)(nil)
-	_ storecontract.TemporalIndexCapability         = (*Store)(nil)
-	_ storecontract.VectorIndexCapability           = (*Store)(nil)
-	_ storecontract.FilteredVectorSearchCapability  = (*Store)(nil)
-	_ storecontract.HighFrequencyIndexCapability    = (*Store)(nil)
+	_ storecontract.Store                                = (*Store)(nil)
+	_ storecontract.Lifecycle                            = (*Store)(nil)
+	_ storecontract.NodeCRUDCapability                   = (*Store)(nil)
+	_ storecontract.RelationshipCRUDCapability           = (*Store)(nil)
+	_ storecontract.AdjacencyCapability                  = (*Store)(nil)
+	_ storecontract.BulkReadCapability                   = (*Store)(nil)
+	_ storecontract.BatchCapability                      = (*Store)(nil)
+	_ storecontract.HistoryCapability                    = (*Store)(nil)
+	_ storecontract.StatsCapability                      = (*Store)(nil)
+	_ storecontract.IterationCapability                  = (*Store)(nil)
+	_ storecontract.NodeIntegrityHashCapability          = (*Store)(nil)
+	_ storecontract.EndpointIntegrityHashCapability      = (*Store)(nil)
+	_ generatedcreate.RelationshipEndpointHashCapability = (*Store)(nil)
+	_ storecontract.PropertyIndexCapability              = (*Store)(nil)
+	_ storecontract.TemporalIndexCapability              = (*Store)(nil)
+	_ storecontract.VectorIndexCapability                = (*Store)(nil)
+	_ storecontract.FilteredVectorSearchCapability       = (*Store)(nil)
+	_ storecontract.HighFrequencyIndexCapability         = (*Store)(nil)
 )